# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/meego/work_item_meetings/chat_auto_bind": {
            "get": {
                "description": "获取群自动关联日程的状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "获取群自动关联日程的状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "查询某个工作项的自动加群状态",
                        "name": "work_item_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetAutoBindCalendarStatusResp"
                        }
                    }
                }
            },
            "post": {
                "description": "群自动关联日程",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "群自动关联日程",
                "parameters": [
                    {
                        "description": "参数",
                        "name": "ChatAutoBindCalendarParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ChatAutoBindCalendarParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrMsgResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ChatAutoBindCalendarParam": {
            "type": "object",
            "properties": {
                "enable": {
                    "description": "功能启用",
                    "type": "boolean"
                },
                "project_key": {
                    "type": "string"
                },
                "work_item_id": {
                    "type": "integer"
                },
                "work_item_type_key": {
                    "type": "string"
                }
            }
        },
        "handler.ErrMsgResp": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "handler.GetAutoBindCalendarStatusResp": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                }
            }
        },
...
    }
}